requirements.txt
altair==5.2.0
anyio==4.1.0
appdirs==1.4.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Automat==0.8.0
Babel==2.13.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
cachetools==5.3.3
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.5
cloud-init==23.3.1
colorama==0.4.3
comm==0.2.0
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.6.1
entrypoints==0.3
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
gitdb==4.0.11
GitPython==3.1.42
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.1.1
incremental==16.10.1
ipykernel==6.27.1
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter-server==2.12.1
jupyter-server-terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
jupyterlab-widgets==3.0.9
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libusb1==3.0.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==4.2.0
mysql-connector-python==8.1.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
ndeflib==0.3.3
nest-asyncio==1.5.8
netifaces==0.10.4
nfcpy==1.0.4
notebook==7.0.6
notebook-shim==0.2.3
numpy==1.24.4
oauthlib==3.1.0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil-resolve-name==1.3.10
platformdirs==3.9.1
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.21.12
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydeck==0.8.1b0
pyDes==2.0.1
pygments==2.17.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.6.3
PyNaCl==1.3.0
pynfc==0.1.1
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==5.3.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.1
requests==2.31.0
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.13.2
SecretStorage==2.3.1
Send2Trash==1.8.2
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.1
sniffio==1.3.0
sos==4.5.6
soupsieve==2.5
ssh-import-id==5.10
stack-data==0.6.3
streamlit==1.31.1
systemd-python==234
tenacity==8.2.3
terminado==0.18.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
traitlets==5.14.0
Twisted==18.9.0
typing-extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
validators==0.22.0
virtualenv==20.24.0
wadllib==1.3.3
watchdog==4.0.0
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.6
widgetsnbextension==4.0.9
zipp==3.17.0
zope.interface==4.7.1
